Bug 1565569 - Remove the old pathfinder integration. r=jrmuizel
Differential Revision: https://phabricator.services.mozilla.com/D37862

[wrupdater] From https://hg.mozilla.org/mozilla-central/rev/784c31031f30e608a33de2d5f654e64f7f0e2a0d
