+++ servo-tidy.toml
@@ -11,7 +11,6 @@ packages = [
     "gl_generator",
     "khronos_api",
     "lazy_static",
-    "log",
     "nix",
     "parking_lot",
     "parking_lot_core",
