+++ webrender/Cargo.toml
@@ -16,7 +16,6 @@ debugger = ["ws", "serde_json", "serde", "image_loader", "base64"]
 capture = ["api/serialize", "ron", "serde", "smallvec/serde"]
 replay = ["api/deserialize", "ron", "serde", "smallvec/serde"]
 display_list_stats = ["api/display_list_stats"]
-pathfinder = ["pathfinder_font_renderer", "pathfinder_gfx_utils", "pathfinder_partitioner", "pathfinder_path_utils"]
 serialize_program = ["serde", "webrender_build/serialize_program"]
 no_static_freetype = []
 
@@ -53,28 +52,6 @@ webrender_build = { version = "0.0.1", path = "../webrender_build" }
 malloc_size_of = { version = "0.0.1", path = "../wr_malloc_size_of", package = "wr_malloc_size_of" }
 ws = { optional = true, version = "0.8" }
 svg_fmt = "0.4"
-
-[dependencies.pathfinder_font_renderer]
-git = "https://github.com/pcwalton/pathfinder"
-branch = "webrender"
-optional = true
-# Uncomment to test FreeType on macOS:
-# features = ["freetype"]
-
-[dependencies.pathfinder_gfx_utils]
-git = "https://github.com/pcwalton/pathfinder"
-branch = "webrender"
-optional = true
-
-[dependencies.pathfinder_partitioner]
-git = "https://github.com/pcwalton/pathfinder"
-branch = "webrender"
-optional = true
-
-[dependencies.pathfinder_path_utils]
-git = "https://github.com/pcwalton/pathfinder"
-branch = "webrender"
-optional = true
 
 [dev-dependencies]
 mozangle = "0.1"
